fastapi/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
passlib
bcrypt
python-jose
python-multipart